separate api types
